beesh: Force defined mode for copies of source files
The bee source file and the package source and patch files
are copied into the bee package, the bee archive directory and into
the build archive with `cp`.

The modes of the source files are more or less undefined. Most notably
this is true for the bee source files themself, because git doesn't store
all file permissions in the repository, just the "x" bit.

Add `--no-preserve=mode` option to `cp` so that the undefined modes are
not inherited. The umask of the processes running bee is already
defined to 022.